#10077: Overload binary le in unary le
