+++ docs/source/ttnn/ttnn/api.rst
@@ -277,7 +277,6 @@ Pointwise Binary
    ttnn/logaddexp2_bw
    ttnn/squared_difference_bw
    ttnn/concat_bw
-   ttnn/binary_le_bw
    ttnn/rsub_bw
    ttnn/bias_gelu_bw
    ttnn/binary_ne_bw
