--- docs/source/ttnn/ttnn/ttnn/binary_le_bw.rst
@@ -1,6 +1,0 @@
-.. _ttnn.binary_le_bw:
-
-ttnn.binary_le_bw
-#################
-
-.. autofunction:: ttnn.binary_le_bw
